+++ movement/src/accelerate.cpp
@@ -35,7 +35,7 @@ void accelerate_linear(Speed from_speed, Speed to_speed, double accel_per_sec, i
     {
         left_accelerator.step();
         right_accelerator.step();
-        MOVEMENT_FUNCTION(left_accelerator.speed(), right_accelerator.speed());
+        MOVEMENT_FUNCTION(static_cast<int>(left_accelerator.speed()), static_cast<int>(right_accelerator.speed()));
         msleep(left_accelerator.get_msleep_time()); // msleep time will be the same for both, so just use left
     }
     MOVEMENT_FUNCTION(to_speed.left, to_speed.right);
@@ -62,7 +62,7 @@ void accelerate_sinusoidal(Speed from_speed, Speed to_speed, double accel_per_se
     {
         left_accelerator.step();
         right_accelerator.step();
-        MOVEMENT_FUNCTION(left_accelerator.speed(), right_accelerator.speed());
+        MOVEMENT_FUNCTION(static_cast<int>(left_accelerator.speed()), static_cast<int>(right_accelerator.speed()));
         msleep(left_accelerator.get_msleep_time()); // msleep time will be the same for both, so just use left
     }
     MOVEMENT_FUNCTION(to_speed.left, to_speed.right);
